--- a/crates/bls48581/Cargo.toml
+++ b/crates/bls48581/Cargo.toml
@@ -12,5 +12,13 @@ hex = "0.4.3"
 serde_json = "1.0.117"
 uniffi = {  version= "0.25", features = ["cli"]}
 
+[dev-dependencies]
+criterion = { version = "0.4", features = ["html_reports"] }
+rand = "0.8.5"
+
 [build-dependencies]
 uniffi = { version = "0.25", features = [ "build" ] }
+
+[[bench]]
+name = "bench_bls"
+harness = false