The final update prior to the v2.0 release.

Extreme performance drilldown on the BLS implementation, needed for fast confirmation times. Also adds some parameterization to connection management for bootstrap systems wanting to expand out the defaults.

## Before

Prior (golang) commits:
|poly_size|time|
|-|-|
|16|43ms|
|128|336ms|
|1024|2808ms|
|65536|165.651s|

Proofs:
|poly_size|time|
|-|-|
|16|43ms|
|128|359ms|
|1024|2.826s|
|65536|178.355s|

Prior (rust) commits in ms:
|poly_size|time|
|-|-|
|16|12ms|
|128|93ms|
|1024|736ms|
|65536|47.521s|

Proofs:
|poly_size|time|
|-|-|
|16|16ms|
|128|133ms|
|1024|1.102s|
|65536|73.836s|

## After
Commits:
|poly_size|time|
|-|-|
|16|166**µ**s|
|128|166**µ**s|
|1024|200ms|
|65536|9.592s|

Proofs:
|poly_size|time|
|-|-|
|16|164**µ**s|
|128|166**µ**s|
|1024|331ms|
|65536|17.402s|

The final amount of juice to squeeze on this would be via parallelization of the DFT –  should be able to bring down the time down on the 65536 case by a strong degree.